requirements.txt
telethon
praw
requests
pillow